refactored extract.py and added more tests
